+++ src/utils.c
@@ -1,10 +1,10 @@
 #include "header.h"
 
 // Linear congruential random number generator
-double rn(unsigned long *seed)
+double rn(unsigned long long *seed)
 {
-  unsigned long a = 19073486328125;
-  unsigned long m = 281474976710656;
+  unsigned long long a = 19073486328125;
+  unsigned long long m = 281474976710656;
 
   *seed = (a*(*seed)) % m;
 
@@ -12,10 +12,10 @@ double rn(unsigned long *seed)
 }
 
 // Linear congruential random number generator for integer in range [min max)
-int rni(unsigned long *seed, int min, int max)
+int rni(unsigned long long *seed, int min, int max)
 {
-  unsigned long a = 19073486328125;
-  unsigned long m = 281474976710656;
+  unsigned long long a = 19073486328125;
+  unsigned long long m = 281474976710656;
 
   *seed = (a*(*seed)) % m;
